+++ upload/upload_test.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/NBISweden/sda-cli/helpers"
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/session"
@@ -41,7 +42,7 @@ func (suite *TestSuite) TestConfigNoFile() {
 	}
 	configPath := "nofile.conf"
 
-	_, err := LoadConfigFile(configPath)
+	_, err := helpers.LoadConfigFile(configPath)
 	assert.EqualError(suite.T(), err, msg)
 }
 
@@ -66,7 +67,7 @@ encrypt = False
 		log.Printf("failed to write temp config file, %v", err)
 	}
 
-	_, err = LoadConfigFile(configPath.Name())
+	_, err = helpers.LoadConfigFile(configPath.Name())
 	assert.EqualError(suite.T(), err, "key-value delimiter not found: guess_mime_type!True\n")
 }
 
@@ -92,7 +93,7 @@ func (suite *TestSuite) TestConfigS3cmdFileFormat() {
 		log.Printf("failed to write temp config file, %v", err)
 	}
 
-	_, err = LoadConfigFile(configPath.Name())
+	_, err = helpers.LoadConfigFile(configPath.Name())
 	assert.NoError(suite.T(), err)
 }
 
@@ -105,7 +106,7 @@ func (suite *TestSuite) TestConfigMissingCredentials() {
 
 	defer os.Remove(configPath.Name())
 
-	_, err = LoadConfigFile(configPath.Name())
+	_, err = helpers.LoadConfigFile(configPath.Name())
 	assert.EqualError(suite.T(), err, "failed to find credentials in configuration file")
 }
 
@@ -125,7 +126,7 @@ access_key = someUser
 		log.Printf("failed to write temp config file, %v", err)
 	}
 
-	_, err = LoadConfigFile(configPath.Name())
+	_, err = helpers.LoadConfigFile(configPath.Name())
 	assert.EqualError(suite.T(), err, "failed to find endpoint in configuration file")
 }
 
@@ -158,7 +159,7 @@ encrypt = False
 		log.Printf("failed to write temp config file, %v", err)
 	}
 
-	_, err = LoadConfigFile(configPath.Name())
+	_, err = helpers.LoadConfigFile(configPath.Name())
 	assert.NoError(suite.T(), err)
 }
 
@@ -218,7 +219,7 @@ func (suite *TestSuite) TestSampleNoFiles() {
 	assert.EqualError(suite.T(), Upload(os.Args), "-config is not a valid target directory")
 
 	// Test uploadFiles function
-	config, _ := LoadConfigFile(configPath.Name())
+	config, _ := helpers.LoadConfigFile(configPath.Name())
 	var files []string
 
 	err = uploadFiles(files, files, "", config)
@@ -229,21 +230,21 @@ func (suite *TestSuite) TestTokenExpiration() {
 	// Token without exp claim
 	// #nosec G101
 	token := "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxNzA3NDgzOTQ0IiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMn0.7r3JJptaxQpuN0I6JwEdfIchf7OOXu--OMFprfMtwzXl2UpmjGVeGy0LWhuzG4LljA2uAp5SPrWzz_U5YKcjuw"
-	expiring, err := CheckTokenExpiration(token)
+	expiring, err := helpers.CheckTokenExpiration(token)
 	assert.EqualError(suite.T(), err, "could not parse token, reason: no expiration date")
 	assert.False(suite.T(), expiring)
 
 	// Token with expired date
 	// #nosec G101
 	token = "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxNzA3NDgzOTQ0IiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMiwiZXhwIjoxNTE2MjM5MDIyfQ.bjYdbKzzR7jbZpLgm_bCqOr_wuaO8KSCEdVJpKEh1pdJ-7klsHdOwCQoBxbmdVPIVHE0jfEEzc9IvtztTeejmg"
-	expiring, err = CheckTokenExpiration(token)
+	expiring, err = helpers.CheckTokenExpiration(token)
 	assert.NoError(suite.T(), err)
 	assert.True(suite.T(), expiring)
 
 	// Token with valid expiration
 	// #nosec G101
 	token = "eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxNzA3NDgzOTQ0IiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWUsImlhdCI6MTUxNjIzOTAyMiwiZXhwIjoxNzA3NDgzOTQ0fQ.D7hrpd3ROXp53NnXa0PL9js2Oi1KqpKpkVMic1B23X84ksX9kbbtn4Ad4BkhO8Tm35a5hBu95CGgw5b06sd3LQ"
-	expiring, err = CheckTokenExpiration(token)
+	expiring, err = helpers.CheckTokenExpiration(token)
 	assert.NoError(suite.T(), err)
 	assert.False(suite.T(), expiring)
 }
